fix: avoid undefined behavior in signed integer 0 bitwise complement
This addresses an ubsan error because 0 is a signed integer. In other places, we avoid this by expliclitly using a `static_cast<size_t>(0)` (another option could be a literal, `0ull`).

+++ src/algorithms/calorimetry/CalorimeterHitReco.cc
@@ -101,7 +101,7 @@ void CalorimeterHitReco::init(const dd4hep::Detector* detector, std::shared_ptr<
         local_mask = id_spec.get_mask(fields);
         // use all fields if nothing provided
         if (fields.empty()) {
-            local_mask = ~0;
+            local_mask = ~static_cast<size_t>(0);
         }
     }
